Health Model/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
pandas
numpy
joblib

pillow
pytesseract
pdf2image
python-multipart